add_screen_for_statistics

+++ app/src/main/java/com/example/myfinanceapplication/model/Cost.kt
@@ -3,19 +3,19 @@ package com.example.myfinanceapplication.model
 data class Cost(
     var costId: String,
     var titleOfCost: String?,
-    var moneyCost: Long,
+    var moneyCost: Double,
     var date: String,
     var category: String?,
     var comment: String?,
     var isExpense: Boolean = true,
 )
 {
-    constructor() : this("", "", 0, "", "", "")
+    constructor() : this("", "", 0.00, "", "", "")
     var goal: String = ""
     constructor(
         costId: String,
         titleOfCost: String?,
-        moneyCost: Long,
+        moneyCost: Double,
         date: String,
         category: String?,
         goal: String,
